+++ src/Utility/AbstractBackupUtility.php
@@ -83,17 +83,16 @@ abstract class AbstractBackupUtility
      *
      * It provides all `getX()` methods to get properties.
      *
-     * @param string $name
-     * @param array $arguments
+     * @param string $method
+     * @param array $args
      * @return mixed
      * @since 2.14.0
-     * @throws \BadMethodCallException With a no existing property or method.
      * @phpstan-ignore missingType.iterableValue
      */
-    public function __call(string $name, array $arguments = []): mixed
+    public function __call(string $method, array $args): mixed
     {
-        if (str_starts_with($name, 'get')) {
-            $property = substr($name, 3);
+        if (str_starts_with($method, 'get')) {
+            $property = substr($method, 3);
             foreach ([$property, lcfirst($property)] as $property) {
                 if (property_exists($this, $property)) {
                     return $this->{$property};
@@ -101,7 +100,7 @@ abstract class AbstractBackupUtility
             }
         }
 
-        throw new BadMethodCallException('Method `' . $this::class . '::' . $name . '()` does not exist.');
+        throw new BadMethodCallException('Method `' . $this::class . '::' . $method . '()` does not exist.');
     }
 
     /**